services/llm-gateway/pyproject.toml
[project]
name = "llm-gateway"
version = "0.1.0"
description = "LLM Gateway service for UTA Travel Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "zhipuai>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"